refactor: Replace `AuthAuthenticator` with `AuthInterceptor` for handling token refresh logic and update related tests

+++ rest/src/main/kotlin/io/vinicius/sak/rest/annotation/Retry.kt
@@ -7,9 +7,9 @@ package io.vinicius.sak.rest.annotation
  * method-level annotation overrides the service-level one, and [@NoRetry][NoRetry] disables retries for a method that
  * would otherwise inherit the service default.
  *
- * Only idempotent verbs (GET, PUT, DELETE) may be retried. A service-level `@Retry` is applied to those methods only and
- * silently skipped for POST/PATCH; a **method-level** `@Retry` or `@NoRetry` on a POST/PATCH method is a compile-time
- * error, preventing accidental request duplication.
+ * Only idempotent verbs (GET, PUT, DELETE) may be retried. A service-level `@Retry` is applied to those methods only
+ * and silently skipped for POST/PATCH; a **method-level** `@Retry` or `@NoRetry` on a POST/PATCH method is a
+ * compile-time error, preventing accidental request duplication.
  *
  * At compile time the `rest-compiler` KSP processor copies the resolved annotation onto the generated `<Name>Retrofit`
  * method, and [io.vinicius.sak.rest.interceptor.RetryInterceptor] reads it at runtime via Retrofit's
@@ -18,7 +18,7 @@ package io.vinicius.sak.rest.annotation
  * Usage:
  * ```kotlin
  * @Service
- * @Retry(maxAttempts = 3, delay = 1) // default for all idempotent methods
+ * @Retry(maxAttempts = 3, delay = 1.0) // default for all idempotent methods
  * interface UserService {
  *     @GET("users/{id}")
  *     @Retry(maxAttempts = 2) // override for this endpoint
@@ -31,11 +31,11 @@ package io.vinicius.sak.rest.annotation
  * ```
  *
  * @param maxAttempts Total number of tries including the first attempt. Must be >= 1.
- * @param delay How long to wait before each retry, in seconds.
+ * @param delay How long to wait before each retry, in seconds. Fractional values are allowed (e.g. `0.5`).
  */
 @Target(AnnotationTarget.CLASS, AnnotationTarget.FUNCTION)
 @Retention(AnnotationRetention.RUNTIME)
 annotation class Retry(
     val maxAttempts: Int = 3,
-    val delay: Long = 1,
+    val delay: Double = 1.0,
 )
